Sync test files from t/nqp repo.

+++ nqp-cc/t/nqp/19-file-ops.t
@@ -4,21 +4,21 @@
 
 plan(40);
 
-ok( nqp::stat('t/nqp/CREDITS', nqp::const::STAT_EXISTS) == 1, 'nqp::stat exists');
+ok( nqp::stat('CREDITS', nqp::const::STAT_EXISTS) == 1, 'nqp::stat exists');
 ok( nqp::stat('AARDVARKS', nqp::const::STAT_EXISTS) == 0, 'nqp::stat not exists');
 
 ok( nqp::stat('t', nqp::const::STAT_ISDIR) == 1, 'nqp::stat is directory');
-ok( nqp::stat('t/nqp/CREDITS', nqp::const::STAT_ISDIR) == 0, 'nqp::stat not directory');
+ok( nqp::stat('CREDITS', nqp::const::STAT_ISDIR) == 0, 'nqp::stat not directory');
 
-ok( nqp::stat('t/nqp/CREDITS', nqp::const::STAT_ISREG) == 1, 'nqp::stat is regular file');
+ok( nqp::stat('CREDITS', nqp::const::STAT_ISREG) == 1, 'nqp::stat is regular file');
 ok( nqp::stat('t', nqp::const::STAT_ISREG) == 0, 'nqp::stat not regular file');
 
-my $credits := nqp::open('t/nqp/CREDITS', 'r');
+my $credits := nqp::open('CREDITS', 'r');
 ok( $credits, 'nqp::open for read');
 ok( nqp::tellfh($credits) == 0, 'nqp::tellfh start of file');
 my $line := nqp::readlinefh($credits);
-ok( nqp::chars($line) == 5 || nqp::chars($line) == 6, 'nqp::readlinefh line to read'); # =pod\n, optionally \r\n on windows
-ok( nqp::tellfh($credits) == 5 || nqp::tellfh($credits) == 6, 'nqp::tellfh line two');
+ok( nqp::chars($line) == 5, 'nqp::readlinefh line to read'); # =pod\n
+ok( nqp::tellfh($credits) == 5, 'nqp::tellfh line two');
 my $rest := nqp::readallfh($credits);
 ok( nqp::chars($rest) > 100, 'nqp::readallfh lines to read');
 ok( nqp::tellfh($credits) == nqp::chars($line) + nqp::chars($rest), 'nqp::tellfh end of file');
@@ -84,34 +84,44 @@ ok(nqp::chars(nqp::readallfh($fh)) == 2, 'switching to ansi results in 2 chars f
 nqp::closefh($fh);
 
 ## chdir
-nqp::chdir('t');
-$fh := nqp::open('../' ~ $test-file, 'r');
-nqp::setencoding($fh, 'utf8');
-ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can chdir into a subdir');
-nqp::closefh($fh);
+if nqp::getcomp('nqp').backend.name eq 'jvm' {
+    ok(1, "$_ # Skipped: chdir is not possible on jvm") for (33, 34, 35);
+}
+else {
+    nqp::chdir('t');
+    $fh := nqp::open('../' ~ $test-file, 'r');
+    nqp::setencoding($fh, 'utf8');
+    ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can chdir into a subdir');
+    nqp::closefh($fh);
 
-nqp::chdir('..');
-$fh := nqp::open($test-file, 'r');
-nqp::setencoding($fh, 'utf8');
-ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can chdir back to the parent dir');
-nqp::closefh($fh);
+    nqp::chdir('..');
+    $fh := nqp::open($test-file, 'r');
+    nqp::setencoding($fh, 'utf8');
+    ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can chdir back to the parent dir');
+    nqp::closefh($fh);
 
-## mkdir
-nqp::mkdir($test-file ~ '-dir', 0o777);
-nqp::chdir($test-file ~ '-dir');
-$fh := nqp::open('../' ~ $test-file, 'r');
-nqp::setencoding($fh, 'utf8');
-ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can create a new directory');
-nqp::closefh($fh);
-nqp::chdir('..');
+    ## mkdir
+    nqp::mkdir($test-file ~ '-dir', 0o777);
+    nqp::chdir($test-file ~ '-dir');
+    $fh := nqp::open('../' ~ $test-file, 'r');
+    nqp::setencoding($fh, 'utf8');
+    ok(nqp::chars(nqp::readallfh($fh)) == 1, 'we can create a new directory');
+    nqp::closefh($fh);
+    nqp::chdir('..');
 
-nqp::rmdir($test-file ~ '-dir');
-nqp::unlink($test-file);
+    nqp::rmdir($test-file ~ '-dir');
+    nqp::unlink($test-file);
+}
 
-$fh := nqp::open('t/nqp/19-readline.txt', 'r');
-ok(nqp::readlinefh($fh) eq "line1\r",   'reading a line till CR');
-ok(nqp::readlinefh($fh) eq "line2\r\n", 'reading a line till CRLF');
-ok(nqp::readlinefh($fh) eq "line3\n",   'reading a line till LF');
-ok(nqp::readlinefh($fh) eq "\n",          'reading an empty line');
-ok(nqp::readlinefh($fh) eq "line4",     'reading a line till EOF');
-nqp::closefh($fh);
+if nqp::getcomp('nqp').backend.name eq 'parrot' {
+    ok(1, "ok $_ # Skipped: readlinefh is broken on parrot") for (36, 37, 38, 39, 40);
+}
+else {
+    $fh := nqp::open('t/nqp/19-readline.txt', 'r');
+    ok(nqp::readlinefh($fh) eq "line1\r",   'reading a line till CR');
+    ok(nqp::readlinefh($fh) eq "line2\r\n", 'reading a line till CRLF');
+    ok(nqp::readlinefh($fh) eq "line3\n",   'reading a line till LF');
+    ok(nqp::readlinefh($fh) eq "\n",          'reading an empty line');
+    ok(nqp::readlinefh($fh) eq "line4",     'reading a line till EOF');
+    nqp::closefh($fh);
+}
